requirements.txt
# Core
pyrogram
tgcrypto-pyrofork
playwright

# Analyze
plotly
pandas

# Utils
python-i18n[YAML]
phonenumbers
dotenv
loguru
tqdm
InquirerPy
pyexiv2
aiohttp
bs4

# Development
pytest
pytest-cov